refactor(store): move NgxsExecutionStrategy symbols into related file
